feat: [PL-23577]: User Deletion flow V2 behind FF USER_DELETION_V2 (#42297)
* feat: [PL-30332]: Fetching pipeline executions data from TS DB

* feat: [PL-29306]: Alerting for GitSyncError per Application level.

* feat: [PL-23577]: Introducing FF

* feat: [PL-23577]: User Deletion flow V2 behind FF USER_DELETION_V2

* feat: [PL-23577]: Removing wrongly added files.

* feat: [PL-23577]: Removing unused imports.

* feat: [PL-28659]: FF name change

* feat: [PL-23577]: Adding a missing line from FF.

* feat: [PL-23577]: Not throwing exception when the user is part of NG. only logging

* feat: [PL-23577]: Passing the user update changes of NG to CG.

* feat: [PL-29306]: Alerting for GitSyncError per Application level.

* feat: [PL-29306]: Alerting for GitSyncError per Application level.

* Rebase develop

* feat: [PL-23577]: Rebasing with develop.

* feat: [PL-23577]: Removing unwanted files.

* feat: [PL-23577]: Addressing review comments.

* feat: [PL-23577]: Addressing review comments.

* feat: [PL-23577]: Addressing review comments.

* feat: [PL-23577]: Changing ownership.

* Resolving conflicts with develop.

* Rebase develop

+++ 400-rest/src/main/java/software/wings/resources/UserResourceNG.java
@@ -17,13 +17,16 @@ import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.annotations.dev.TargetModule;
 import io.harness.beans.FeatureFlag;
+import io.harness.beans.FeatureName;
 import io.harness.beans.PageRequest;
 import io.harness.beans.PageResponse;
 import io.harness.exception.InvalidRequestException;
 import io.harness.exception.UnauthorizedException;
 import io.harness.exception.WingsException;
+import io.harness.ff.FeatureFlagService;
 import io.harness.mappers.AccountMapper;
 import io.harness.ng.core.dto.UserInviteDTO;
+import io.harness.ng.core.user.NGRemoveUserFilter;
 import io.harness.ng.core.user.PasswordChangeDTO;
 import io.harness.ng.core.user.PasswordChangeResponse;
 import io.harness.ng.core.user.TwoFactorAdminOverrideSettings;
@@ -46,6 +49,7 @@ import software.wings.beans.UserInvite;
 import software.wings.security.authentication.TwoFactorAuthenticationManager;
 import software.wings.security.authentication.TwoFactorAuthenticationMechanism;
 import software.wings.security.authentication.TwoFactorAuthenticationSettings;
+import software.wings.service.impl.UserServiceHelper;
 import software.wings.service.intfc.AccountService;
 import software.wings.service.intfc.SignupService;
 import software.wings.service.intfc.UserService;
@@ -90,11 +94,14 @@ import retrofit2.http.Body;
 public class UserResourceNG {
   private final UserService userService;
   private final SignupService signupService;
+  private final UserServiceHelper userServiceHelper;
   private final TwoFactorAuthenticationManager twoFactorAuthenticationManager;
   private final AccountService accountService;
   private final ScimUserService scimUserService;
   private static final String COMMUNITY_ACCOUNT_EXISTS = "A community account already exists";
   private static final String ACCOUNT_ADMINISTRATOR_USER_GROUP = "Account Administrator";
+
+  @Inject private FeatureFlagService featureFlagService;
 
   @POST
   public RestResponse<UserInfo> createNewUserAndSignIn(UserRequestDTO userRequest) {
@@ -187,7 +194,16 @@ public class UserResourceNG {
   @DELETE
   public RestResponse<Boolean> deleteUser(
       @QueryParam("accountId") String accountId, @QueryParam("userId") String userId) {
-    userService.delete(accountId, userId);
+    if (featureFlagService.isEnabled(FeatureName.PL_USER_DELETION_V2, accountId)) {
+      userServiceHelper.deleteUserFromNG(userId, accountId, NGRemoveUserFilter.ACCOUNT_LAST_ADMIN_CHECK);
+      if (!userService.isUserPartOfAnyUserGroupInCG(userId, accountId)) {
+        log.warn("User {}, is being deleted from CG, since he is not part of any user-groups in CG",
+            userService.get(userId).getEmail());
+        userService.delete(accountId, userId);
+      }
+    } else {
+      userService.delete(accountId, userId);
+    }
     return new RestResponse<>(true);
   }
 
@@ -203,6 +219,9 @@ public class UserResourceNG {
   @Path("/scim/disabled")
   public RestResponse<Boolean> disableScimUser(@QueryParam("accountId") String accountId,
       @QueryParam("userId") String userId, @QueryParam("disabled") boolean disabled) {
+    if (featureFlagService.isEnabled(FeatureName.PL_USER_DELETION_V2, accountId)) {
+      disabled = false;
+    }
     return new RestResponse<>(scimUserService.changeScimUserDisabled(accountId, userId, disabled));
   }
 
@@ -211,6 +230,13 @@ public class UserResourceNG {
   public RestResponse<ScimUser> patchUpdateScimUser(
       @QueryParam("accountId") String accountId, @QueryParam("userId") String userId, @Body PatchRequest patchRequest) {
     return new RestResponse<>(scimUserService.updateUser(accountId, userId, patchRequest));
+  }
+
+  @PUT
+  @Path("/scim/patch/details")
+  public RestResponse<ScimUser> patchUpdateScimUserDetails(
+      @QueryParam("accountId") String accountId, @QueryParam("userId") String userId, @Body PatchRequest patchRequest) {
+    return new RestResponse<>(scimUserService.updateUserDetails(accountId, userId, patchRequest));
   }
 
   @PUT
